split examples between "provide" side and "fetch" side
